Moved cli commands' implementations to separate files
