+++ packages/discovery/src/cli.ts
@@ -1,19 +1,10 @@
-import { assert, Logger } from '@l2beat/backend-tools'
-import { providers } from 'ethers'
+import { Logger } from '@l2beat/backend-tools'
 
+import { discoverCommand } from './cli/discoverCommand'
 import { handleCli } from './cli/handleCli'
-import {
-  DiscoveryCliConfig,
-  DiscoveryModuleConfig,
-  getDiscoveryCliConfig,
-} from './config/config.discovery'
-import { ConfigReader } from './discovery/config/ConfigReader'
-import { dryRunDiscovery, runDiscovery } from './discovery/runDiscovery'
-import { runInversion } from './inversion/runInversion'
-import { singleDiscoveryCommand } from './singleDiscovery'
-import { ChainId } from './utils/ChainId'
-import { EtherscanLikeClient } from './utils/EtherscanLikeClient'
-import { HttpClient } from './utils/HttpClient'
+import { invertCommand } from './cli/invertCommand'
+import { singleDiscoveryCommand } from './cli/singleDiscoveryCommand'
+import { getDiscoveryCliConfig } from './config/config.discovery'
 
 main().catch((e) => {
   console.error(e)
@@ -28,92 +19,4 @@ async function main(): Promise<void> {
   await discoverCommand(config, logger)
   await invertCommand(config, logger)
   await singleDiscoveryCommand(config, logger)
-}
-
-export async function discover(
-  config: DiscoveryModuleConfig,
-  logger: Logger = Logger.DEBUG,
-): Promise<void> {
-  const cliConfig = getDiscoveryCliConfig({
-    mode: 'discover',
-    project: config.project,
-    chain: config.chainId,
-    dryRun: config.dryRun === true,
-    dev: config.dev === true,
-    sourcesFolder: config.sourcesFolder,
-    discoveryFilename: config.discoveryFilename,
-    blockNumber: config.blockNumber,
-  })
-
-  await discoverCommand(cliConfig, logger)
-}
-
-async function discoverCommand(
-  config: DiscoveryCliConfig,
-  logger: Logger,
-): Promise<void> {
-  if (!config.discovery) {
-    return
-  }
-  const discoverConfig = config.discovery
-  const chainConfig = config.chain
-
-  assert(
-    chainConfig.chainId === discoverConfig.chainId,
-    'Chain config does not match discovery config! Update "discovery.config" file or config.json of your project',
-  )
-
-  const http = new HttpClient()
-  const provider = new providers.StaticJsonRpcProvider(chainConfig.rpcUrl)
-  const etherscanClient = EtherscanLikeClient.createForDiscovery(
-    http,
-    chainConfig.etherscanUrl,
-    chainConfig.etherscanApiKey,
-    chainConfig.etherscanUnsupported,
-  )
-  const configReader = new ConfigReader()
-
-  if (discoverConfig.dryRun) {
-    logger = logger.for('DryRun')
-    logger.info('Starting')
-
-    await dryRunDiscovery(
-      provider,
-      etherscanClient,
-      config.chain.multicall,
-      configReader,
-      discoverConfig,
-    )
-    return
-  }
-
-  logger = logger.for('Discovery')
-  logger.info('Starting discovery...\n')
-  logger.info(`Project: ${discoverConfig.project}`)
-  logger.info(`Chain: ${ChainId.getName(discoverConfig.chainId)}\n`)
-  await runDiscovery(
-    provider,
-    etherscanClient,
-    config.chain.multicall,
-    configReader,
-    discoverConfig,
-  )
-}
-
-async function invertCommand(
-  config: DiscoveryCliConfig,
-  logger: Logger,
-): Promise<void> {
-  if (!config.invert) {
-    return
-  }
-
-  const { project, useMermaidMarkup, chainId } = config.invert
-
-  const configReader = new ConfigReader()
-
-  logger = logger.for('Inversion')
-  logger.info('Starting')
-
-  await runInversion(project, configReader, useMermaidMarkup, chainId)
 }
