+++ packages/discovery/src/index.ts
@@ -1,4 +1,4 @@
-export { discover } from './cli'
+export { discover } from './cli/discoverCommand'
 export { AddressAnalyzer } from './discovery/analysis/AddressAnalyzer'
 export { ConfigReader } from './discovery/config/ConfigReader'
 export { DiscoveryConfig } from './discovery/config/DiscoveryConfig'
